+++ TiltSwitch/AppMetadata.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum AppMetadata {
+    static let displayName = "TiltSwitch"
+    static let statusItemTitle = "TiltSwitch"
+    static let websiteURL = URL(string: "https://boblebol.github.io/TiltSwitch/")!
+    static let githubURL = URL(string: "https://github.com/Boblebol/TiltSwitch")!
+    static let cameraSettingsURL = URL(string: "x-apple.systempreferences:com.apple.preference.security?Privacy_Camera")!
+    static let keyboardShortcutsSettingsURL = URL(string: "x-apple.systempreferences:com.apple.preference.keyboard?Shortcuts")!
+}